init() moved to redis verticle
