+++ vertx/example-configs/config-depl.json
@@ -29,15 +29,13 @@
 			"redisPassword": "abcd",
 			"redisMaxWaitingHandlers": 5000,
 			"redisHost": "redis-redis-cluster.redis.svc.cluster.local",
-			"redisPort": 6379
+			"redisPort": 6379,
+			"catalogueHost": "",
+			"cataloguePort": 443
 		},
 		{
 			"id": "iudx.ingestion.pipeline.processor.ProcessorVerticle",
-			"verticleInstances": 1,
-			"catalogueHost": "",
-			"cataloguePort": 443,
-			"keystore": "",
-			"keystorePassword": ""
+			"verticleInstances": 1
 		}
 	]
 }
