+++ vertx/example-configs/config-dev.json
@@ -28,15 +28,13 @@
 			"redisMaxWaitingHandlers": 1024,
 			"redisPoolRecycleTimeout": 1500,
 			"redisHost": "redis-redis-cluster.redis.svc.cluster.local",
-			"redisPort": 6379
+			"redisPort": 6379,
+			"catalogueHost": "",
+			"cataloguePort": 443
 		},
 		{
 			"id": "iudx.ingestion.pipeline.processor.ProcessorVerticle",
-			"verticleInstances": 1,
-            "catalogueHost": "",
-            "cataloguePort": 443,
-            "keystore": "",
-            "keystorePassword": ""
+			"verticleInstances": 1
 		}
 	]
 }
